requirements.txt
apf_base
python-snappy